[[WIP]]: move slave into details
