+++ node/include/cocaine/detail/service/node/slave.hpp
@@ -24,6 +24,7 @@
 #include "cocaine/service/node/profile.hpp"
 #include "cocaine/service/node/slave/error.hpp"
 #include "cocaine/service/node/slot.hpp"
+#include "cocaine/service/node/slave/channel.hpp"
 
 namespace cocaine {
 
@@ -55,16 +56,6 @@ namespace slave {
 
 using service::node::slave::id_t;
 using service::node::slave::stats_t;
-
-struct channel_t {
-    /// Event to be processed.
-    app::event_t event;
-
-    std::shared_ptr<client_rpc_dispatch_t> dispatch;
-
-    /// An RX stream provided from user. The slave will call its callbacks on every incoming event.
-    std::shared_ptr<api::stream_t> downstream;
-};
 
 } // namespace slave
