+++ node/src/node/slave/fetcher.cpp
@@ -2,7 +2,7 @@
 
 #include <blackhole/logger.hpp>
 
-#include "cocaine/service/node/slave.hpp"
+#include "cocaine/detail/service/node/slave.hpp"
 #include "cocaine/detail/service/node/slave/machine.hpp"
 
 namespace ph = std::placeholders;
