+++ node/src/node/slave/state/spawning.cpp
@@ -8,7 +8,7 @@
 
 #include <cocaine/rpc/actor.hpp>
 
-#include "cocaine/service/node/slave.hpp"
+#include "cocaine/detail/service/node/slave.hpp"
 #include "cocaine/detail/service/node/slave/machine.hpp"
 
 #include "cocaine/detail/service/node/slave/state/handshaking.hpp"
